refactor: consolidate memory operations into dedicated modules with improved filtering and metadata handling
